Move sgf to directive and label parsing to server
1) All the WGo instantiation and so forth used to be handled on the kifu
controller, which was dumb. Now it uses the SGF directive, which is
nicely componentized and has a clear API.

2) Move label parsing to the server. Since we're already parsing
comments using Markdown on the server, there's no reason to offload this
task to the client. It also simplifies things to parse Markdown rather
than HTML, since we don't have to worry about <h1> tags, which look like
coordinate labels.