Add npm-shrinkwrap, closes #58
I discovered that, by default, `npm shrinkwrap` does not include dev
dependencies, where the idea is that you use normal semver dependency
management.

Since we're doing this to try to eliminate some of the errors that
spring up without my knowing it as project creator when someone else
tries to contribute to gokibitz, I added dev dependencies using
`npm shrinkwrap --dev`.

Contributing to someone else's project as a volunteer should be as
painless as possible! I want to be the person encountering and dealing
with breaking changes.

Thanks to @tristil for creating this issue and making me learn about
`npm shrinkwrap`.