Remove JS-powered nav links
JS links are the worst. Every time you cmd + click to open something in
a new tab, it just redirects in the current tab, which is rage-inducing.
If I had anyone to swear at other than myself, I would have done so
numerous times.

The "goToAndCollapse" thing must have come from an example, so it's
possible there's a reason for it, which may mean in the future I need to
recreate some aspect of it, but for now it's gone.