Refactor gkCommentBox to remove isolated scope
Isolated scope works really well in directives, but you can only have
ONE directive with isolated scope on an element at a time. @levelonedev
ran into this when trying to use a third-party directive we found for
notifications on a comment's textarea.

gkCommentBox didn't really rely on having an isolated scope, but it
needs custom functions and variables to be able to be passed into it.
Angular's isolate scope takes care of that in a really nice way, but
it's possible to do separately, using the passed-in attributes object.
You just have to use $parse and $scope.$eval on functions and strings,
respectively, to be able to use them.

It's a little less pretty: hopefully this will be smoothed out in the
directive reworking that's part of Angular 2.0.