Fix rank validation for profile, fix #93
This was a regression caused by a bit of refactoring that was part of
introducing the JS style checker. I had to alter my (too) fancy do while
loop, and I accidentally broke it. Since I quickly checked it against an
SDK rank and it still worked, I missed the fact that it was now grabbing
the wrong suffix. (An off-by-one error!)

This is why we need unit tests.

Huge thanks to @bigblindbais for reporting this regression.

Fix error revealed by our very-helpful linter.