Fix order of getUser / getTotal
The getTotal function was wrongly placed before getUser in the async
series order, meaning getTotal always returned the total number of
comments, not the total number of comments for a given user. That means
that chunkify had to go through ALL comments every time if it was
fetching a list for a user that didn't have enough comments to fill the
request.

This wasn't that big of a deal for a while, but as the number of
comments have grown, it was becoming an increasing drain on CPU.