Create new mention directive
With the node upgrade, I couldn't get ment.io working at all. The
project hasn't been updated in a few years, so I finally decided to stop
trying to fix a third-party module and just build it myself. This made
it easy to give it a little more of a GoKibitz look and feel, so it
wasn't too bad.