Fix unhandled error problem in browserify task
It looks like bulkify was the problem, the tool I was using to require a
bunch of files from a directory. I'm not sure why, why I guess it wasn't
dealing with parsing errors properly, and was exploding.

IT WAS EXASPERATING.

And I will never get those hours of my life back.

But `require-globify` is a better tool than bulkify was, and does
exactly what I want, so that's great.

I guess the lesson here is to think about all of the middlemen involved
in the browserify bundle, and realize that if any of them isn't doing
its part properly, they could be the cause of the problem.