Use content-disposition module to fix filename bug
Filenames for SGFs are created by using the player names. It looks like
when the player names contained some non-ASCII characters, it was
breaking the content-disposition header on the download, which made node
crash.

There was a content-disposition module on NPM apparently designed to
solve this exact problem.

Fixes #137 and #142