Fix problem with double-starring
When you star a comment, it suddenly displays two, rather than one.

This problem was caused by an inadvertent doubling up of the way comment
stars get updated... I'd added star addition and removal to the
toggleStar method before I added a socket listener to react to stars, so
the list was getting updated from two places.

This change lets the socket listener handle it on its own.