Eliminate redundant notifications involving mentions
Fix #114

1. Check the list of users who get notified for being mentioned when a
comment is made, then remove them from the list of users who will be
notified because they have already commented on a move. This reduces
double notifications for people who have commented on a move, then get
mentioned in a subsequent comment.

2. When editing a comment, check the list of users who have already been
notified for being mentioned, and remove them from the list of those
mentioned, so that only *newly* mentioned users get notifications.

Big thanks to @levelonedev for his original implementation of this
feature: it's turned out to be a really popular and important part of
the site.