Fix watcher for comment text
Change it so that it fires for value changes that happen
programmatically, not just changes to the input.

This fixes the problem with a preview only being triggered by typing in
the textarea, and not updating when a user has input a variation through
the widget.

See this post for a very helpful explanation of the difference between
$watch and viewChangeListeners:
https://www.bennadel.com/blog/
2955-watching-and-responding-to-ngmodel-changes-in-angularjs.htm