Check to see if there's a comment for a move before fetching comments
This closes #120. (At last. I think.)

Weirdly, I suddenly realized that I'm _already_ keeping track of what
moves have comments for them as part of the logic that lets us navigate
between moves with comments. (It's stored as `$scope.uniqComments` in
the kifu controller.)

Once I realized that, it's perfectly easy to check whether we need to
actually list comments or not.

Now, the right comment area will say "No comments yet" unless there are
actual comments to be fetched. If there are, then that area will be
blank until comments are actually fetched. (I tried having an animated
throbber there to make it more obvious, but I think it's more
distracting than useful. This way, just the absence of "No comments yet"
can indicate to the quickly moving user that there are comments to wait
for.)